fix: separated Form from search page
